Searchbar Centralization for Admin and Therapist
