+++ frontend/src/components/Admin/Appointment/Appointment.js
@@ -1,9 +1,7 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import BellIcon from "../../images/SVG_files/BellIcon.svg";
-import SearchBar from "../../images/SVG_files/SearchBar.svg";
 import { BASE_ADMIN } from "../../../utils/constants";
-import { useNavigate } from 'react-router-dom'
+import SearchBar from "../SearchBar";
 
 const Appointment = () => {
   const [appointmentsList, setAppointmentsList] = useState([]);
@@ -28,14 +26,6 @@ const Appointment = () => {
         console.error("Error fetching appointments:", error);
       });
   }, []);
-
-  const navigate = useNavigate()
-
-  const logout = () => {
-    localStorage.clear("adminToken");
-    navigate('/admin-login')
-
-  }
 
   const cancelledStyle = { color: "#B00202" };
   const completedStyle = { color: "#02B04A" };
@@ -84,23 +74,7 @@ const Appointment = () => {
   return (
     <div className="w-full m-10 ">
       {/* Search Bar */}
-      <div className="flex w-[95%] justify-between gap-10 items-center">
-        <div className="flex items-center bg-white w-[80%]   pl-4 rounded-lg -[#B4F0FF] ">
-          <img src={SearchBar} alt="SearchBar" />
-
-          <img src={SearchBar} alt="SearchBar" />
-
-          <input
-            type="text"
-            placeholder="Search"
-            className="h-12 ml-5 rounded-lg outline-none w-[100%]"
-          />
-        </div>
-        <div className="p-2 cursor-pointer rounded-full ">
-          <img src={BellIcon} alt=" BellIcon " />
-        </div>
-        <button onClick={logout} className="bg-[#0190B1] px-4 py-2 rounded-md text-white">Logout</button>
-      </div>
+      <SearchBar />
       {/* heading */}
       <div className="w-[95%] flex justify-between items-center text-2xl my-8">
         <h1>Appointments</h1>
