+++ frontend/src/components/Admin/Therapist/Therapist.js
@@ -1,9 +1,7 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import BellIcon from "../../images/SVG_files/BellIcon.svg";
-import SearchBar from "../../images/SVG_files/SearchBar.svg";
 import { BASE_ADMIN } from "../../../utils/constants";
-import { useNavigate } from "react-router-dom";
+import SearchBar from "../SearchBar";
 
 const Therapist = () => {
   const [therapistsDetails, setTherapistsDetails] = useState([]);
@@ -43,14 +41,6 @@ const Therapist = () => {
     fetchTherapistsDetails();
   }, []);
 
-  const navigate = useNavigate()
-
-  const logout = () => {
-    localStorage.clear("adminToken");
-    navigate('/admin-login')
-
-  }
-
   const formatDate = (dateString) => {
     const options = { day: "2-digit", month: "2-digit", year: "numeric" };
     return new Date(dateString).toLocaleDateString("en-GB", options);
@@ -58,23 +48,7 @@ const Therapist = () => {
 
   return (
     <div className="w-full m-10 ">
-      <div className="flex justify-between w-[95%] gap-10 items-center">
-        <div className="flex items-center bg-white w-[80%] border  pl-4 rounded-lg border-[#B4F0FF] ">
-          <img src={SearchBar} alt="SearchBar" />
-
-         <img src={SearchBar} alt="SearchBar"/>
-
-          <input
-            type="text"
-            placeholder="Search"
-            className="h-12 ml-5 rounded-lg outline-none w-[100%]"
-          />
-        </div>
-        <div className="p-2 cursor-pointer rounded-full ">
-          <img src={BellIcon} alt=" BellIcon " />
-        </div>
-        <button onClick={logout} className="bg-[#0190B1] px-4 py-2 rounded-md text-white">Logout</button>
-      </div>
+      <SearchBar />
       <div className="flex w-[95%] justify-between items-center">
         <h1 className="text-2xl  my-8">Therapists</h1>
         <button className="rounded-md py-2 px-4 cursor-pointer  text-lg bg-[#00677F] text-white">
