+++ frontend/src/components/Admin/Transactions/Transactions.js
@@ -1,11 +1,7 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import SearchBar from "../../images/SVG_files/SearchBar.svg";
-import BellIcon from "../../images/SVG_files/BellIcon.svg";
 import { BASE_ADMIN } from "../../../utils/constants";
-import { useNavigate } from "react-router-dom";
-
-
+import SearchBar from "../SearchBar";
 
 const Transactions = () => {
   const [transactions, setTransactions] = useState([])
@@ -28,13 +24,6 @@ const Transactions = () => {
       });
   }, []);
 
-  const navigate = useNavigate()
-
-  const logout = () => {
-    localStorage.clear("adminToken");
-    navigate('/admin-login')
-
-  }
   const cancelledStyle = { color: "#B00202" };
   const completedStyle = { color: "#02B04A" };
 
@@ -64,23 +53,7 @@ const Transactions = () => {
   return (
     <div className="w-full m-10 ">
       {/* Search Bar */}
-      <div className="flex justify-between gap-10 items-center">
-        <div className="flex items-center bg-white w-[80%]  pl-4 rounded-lg-[#B4F0FF] ">
-          <img src={SearchBar} alt="SearchBar" />
-
-          <img src={SearchBar} alt="SearchBar" />
-
-          <input
-            type="text"
-            placeholder="Search"
-            className="h-12 ml-5 rounded-lg outline-none w-[100%]"
-          />
-        </div>
-        <div className="p-2 rounded-full ">
-          <img src={BellIcon} alt=" BellIcon " />
-        </div>
-        <button onClick={logout} className="bg-[#0190B1] px-4 py-2 rounded-md text-white">Logout</button>
-      </div>
+      <SearchBar />
       <div className="flex w-[90%] justify-between items-center">
         <h1 className="text-2xl  my-8">Transactions</h1>
       </div>
